Refactor EntryListWidget into two classes
